+++ apps/bilan-carbone/src/services/serverFunctions/importEmissionFactors.ts
@@ -6,8 +6,16 @@ import { LocaleType } from '@/i18n/config'
 import { getLocale } from '@/i18n/locale'
 import { environmentSubPostsMapping } from '@/services/posts'
 import {
+  COLUMNS,
+  ImportEmissionFactorsResult,
+  ImportError,
+  ParsedRow,
+  ParseResult,
+  PreviewEmissionFactorsResult,
+  PreviewRow,
+} from '@/types/importEmissionFactors.types'
+import {
   buildPostsAndSubPostsCell,
-  getBcTranslations,
   getUnitLabel,
   mapBaseLabelFromTranslations,
   mapQualityLabelFromTranslations,
@@ -17,7 +25,8 @@ import {
 import { parseNumericValue } from '@/utils/number'
 import { flattenSubposts } from '@/utils/post'
 import { withServerResponse } from '@/utils/serverResponse'
-import { EmissionFactorBase, EmissionFactorStatus, Environment, Import, SubPost, Unit } from '@repo/db-common/enums'
+import { getBcTranslations } from '@/utils/translation.utils'
+import { EmissionFactorStatus, Environment, Import, SubPost, Unit } from '@repo/db-common/enums'
 import xlsx from 'node-xlsx'
 import { auth } from '../auth'
 import { NOT_AUTHORIZED } from '../permissions/check'
@@ -25,84 +34,6 @@ import { canCreateEmissionFactor } from '../permissions/emissionFactor.server'
 import { EmissionFactorCommandValidation } from './emissionFactor.command'
 import { prepareExcel } from './file'
 import { getFileUrlFromBucket } from './scaleway'
-
-type ImportError = {
-  line: number
-  key: string
-  value?: string
-}
-
-type ImportEmissionFactorsResult = { success: true; count: number } | { success: false; errors: ImportError[] }
-
-type PreviewRow = {
-  name: string
-  source: string
-  unit: string
-  totalCo2: number
-  postsAndSubPosts: string
-}
-
-type PreviewEmissionFactorsResult = { success: true; rows: PreviewRow[] } | { success: false; errors: ImportError[] }
-
-const COLUMNS = {
-  name: 0,
-  attribute: 1,
-  location: 2,
-  source: 3,
-  unit: 4,
-  customUnit: 5,
-  totalCo2: 6,
-  co2f: 7,
-  ch4f: 8,
-  ch4b: 9,
-  n2o: 10,
-  co2b: 11,
-  sf6: 12,
-  hfc: 13,
-  pfc: 14,
-  otherGES: 15,
-  reliability: 16,
-  technicalRepresentativeness: 17,
-  geographicRepresentativeness: 18,
-  temporalRepresentativeness: 19,
-  completeness: 20,
-  postsAndSubPosts: 21,
-  base: 22,
-  comment: 23,
-} as const
-
-type ParsedRow = {
-  name: string
-  attribute: string | undefined
-  location: string | undefined
-  source: string
-  unit: Unit
-  customUnit: string | null
-  totalCo2: number
-  co2f: number | undefined
-  ch4f: number | undefined
-  ch4b: number | undefined
-  n2o: number | undefined
-  co2b: number | undefined
-  sf6: number | undefined
-  hfc: number | undefined
-  pfc: number | undefined
-  otherGES: number | undefined
-  reliability: number
-  technicalRepresentativeness: number
-  geographicRepresentativeness: number
-  temporalRepresentativeness: number
-  completeness: number
-  subPosts: Record<string, SubPost[]>
-  comment: string | undefined
-  isMonetary: boolean
-  parts: []
-  base: EmissionFactorBase | null
-  rawPostsAndSubPosts: string
-  rawUnit: string
-}
-
-type ParseResult = { success: true; rows: ParsedRow[] } | { success: false; errors: ImportError[] }
 
 async function checkAuth() {
   const session = await auth()
